requirements.txt
# Python Dependencies
faker==24.0.0
pandas==2.2.0
numpy==1.26.3
duckdb==0.10.0
sqlalchemy==2.0.25
great-expectations==0.18.8
requests==2.31.0
python-dotenv==1.0.0
pyarrow==15.0.0
